+++ xmm/base.py
@@ -1,9 +1,10 @@
-from xmm.logger import logger
+import logging
 from xmm.config import conf
+from xmm.logger import ClassPrefixAdapter
 
 
 class Base(object):
 
     def __init__(self):
         self.conf = conf
-        self.logger = logger
+        self.logger = ClassPrefixAdapter(prefix=self.__class__.__name__, logger=logging.getLogger(__name__))