+++ xmm/logger.py
@@ -1,3 +1,12 @@
 import logging
 
-logger = logging.getLogger(__name__)
+
+class ClassPrefixAdapter(logging.LoggerAdapter):
+    def __init__(self, prefix, logger, extra=None):
+        super(ClassPrefixAdapter, self).__init__(logger, extra or {})
+        self.prefix = prefix
+    """
+    Custom Adapter to insert class name to provide context
+    """
+    def process(self, msg, kwargs):
+        return '[%s] %s' % (self.prefix, msg), kwargs